requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
django-decouple==2.1
django-env==0.2.1
django-stubs==5.2.7
django-stubs-ext==5.2.7
graphql-core==3.2.6
lia-web==0.2.3
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
six==1.17.0
sqlparse==0.5.3
strawberry-django-auth==0.378.4
strawberry-graphql==0.284.1
strawberry-graphql-django==0.67.0
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
